src/requirements.txt
Django==1.6.5
django-ipware==0.0.8
wsgiref==0.1.2
